decompose nav menu
