+++ apps/web-main/src/components/ui/Menu/index.ts
@@ -1,0 +1,1 @@
+export * from './AppNavigationMenu';
